pyproject.toml
[project]
name = "archex"
version = "0.5.0"
description = "Architecture extraction & codebase intelligence for the agentic era"
readme = "README.md"
license = { text = "Apache-2.0" }
requires-python = ">=3.11"
authors = [{ name = "Tom" }]
keywords = ["architecture", "codebase", "retrieval", "rag", "ast", "code-intelligence"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Topic :: Software Development :: Libraries",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
]
dependencies = [
    "tree-sitter>=0.23",
    "tree-sitter-python>=0.23",
    "tree-sitter-javascript>=0.23",
    "tree-sitter-typescript>=0.23",
    "tree-sitter-go>=0.23",
    "tree-sitter-rust>=0.23",
    "tree-sitter-java>=0.23",
    "tree-sitter-kotlin>=1.1",
    "tree-sitter-c-sharp>=0.23",
    "tiktoken>=0.7",
    "pydantic>=2.7",
    "networkx>=3.3",
    "click>=8.1",
    "pyyaml>=6.0",
    "fastapi>=0.115",
    "uvicorn[standard]>=0.34",
]

[project.optional-dependencies]
vector = ["onnxruntime>=1.17", "tokenizers>=0.15"]
vector-fast = [
    "fastembed>=0.4",
]
vector-torch = [
    "sentence-transformers>=2.6",
]
openai = ["openai>=1.0"]
anthropic = ["anthropic>=0.30"]
mcp = ["mcp>=1.0"]
langchain = ["langchain-core>=0.2"]
llamaindex = ["llama-index-core>=0.10"]
lsap = ["lsp-client>=0.3; python_version >= '3.12'"]
language-pack = ["tree-sitter-language-pack>=0.13"]
all = ["archex[vector,openai,anthropic,mcp,langchain,llamaindex,language-pack]"]
web = [
    "fastapi>=0.115",
    "uvicorn[standard]>=0.34",
]
dev = [
    "pytest>=8.0",
    "pytest-cov>=5.0",
    "pytest-asyncio>=0.23",
    "ruff>=0.5",
    "pyright>=1.1",
    "pre-commit>=3.7",
    "httpx>=0.27",
]

[project.scripts]
archex = "archex.cli.main:cli"

[project.entry-points."archex.language_adapters"]
# Example: java = "mypackage.adapters:JavaAdapter"

[project.entry-points."archex.pattern_detectors"]
# Example: my_pattern = "mypackage.patterns:detect_my_pattern"

[project.entry-points."archex.benchmark_strategies"]
# Example: custom_strategy = "mypackage.strategies:run_custom"

[project.entry-points."archex.embedders"]
# Example: custom_embedder = "mypackage.embedders:my_embedder_factory"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/archex"]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "UP", "B", "SIM", "TCH"]

[tool.pyright]
pythonVersion = "3.11"
typeCheckingMode = "strict"
include = ["src/archex", "tests"]
exclude = ["tests/fixtures/", ".venv/"]
venvPath = "."
venv = ".venv"

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "--cov=archex --cov-report=term-missing -m 'not slow'"
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
]

[dependency-groups]
dev = [
    "langchain-core>=0.2",
    "llama-index-core>=0.10",
    "mcp>=1.0",
    "numpy>=2.4.2",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "pytest-cov>=7.0.0",
    "scipy>=1.17.1",
    "tree-sitter-language-pack>=0.13",
]
